Doc and code style pass

+++ src/main/java/com/dreammaster/TwilightForest/TF_Loot_Chests.java
@@ -21,52 +21,52 @@ public class TF_Loot_Chests {
     public static void init() {
 
         // Hill1 = small
-        TFTreasure.hill1.common.add(MaterialLibAPI.getStack(Materials.Thaumium, Shapes.dustTiny, (int) (2L)), 6);
-        TFTreasure.hill1.common.add(MaterialLibAPI.getStack(Materials.Bronze, Shapes.ingot, (int) (4L)), 12);
-        TFTreasure.hill1.common.add(MaterialLibAPI.getStack(Materials.Nickel, Shapes.ingot, (int) (4L)), 12);
-        TFTreasure.hill1.common.add(MaterialLibAPI.getStack(Materials.Thaumium, Shapes.dustSmall, (int) (2L)), 6);
-        TFTreasure.hill1.uncommon.add(MaterialLibAPI.getStack(Materials.Manganese, Shapes.ingot, (int) (4L)), 12);
-        TFTreasure.hill1.uncommon.add(MaterialLibAPI.getStack(Materials.Antimony, Shapes.ingot, (int) (4L)), 12);
-        TFTreasure.hill1.rare.add(MaterialLibAPI.getStack(Materials.Thaumium, Shapes.dust, (int) (1L)), 6);
-        TFTreasure.hill1.ultrarare.add(MaterialLibAPI.getStack(Materials.Aluminium, Shapes.dustSmall, (int) (4L)), 12);
+        TFTreasure.hill1.common.add(MaterialLibAPI.getStack(Materials.Thaumium, Shapes.dustTiny, 2), 6);
+        TFTreasure.hill1.common.add(MaterialLibAPI.getStack(Materials.Bronze, Shapes.ingot, 4), 12);
+        TFTreasure.hill1.common.add(MaterialLibAPI.getStack(Materials.Nickel, Shapes.ingot, 4), 12);
+        TFTreasure.hill1.common.add(MaterialLibAPI.getStack(Materials.Thaumium, Shapes.dustSmall, 2), 6);
+        TFTreasure.hill1.uncommon.add(MaterialLibAPI.getStack(Materials.Manganese, Shapes.ingot, 4), 12);
+        TFTreasure.hill1.uncommon.add(MaterialLibAPI.getStack(Materials.Antimony, Shapes.ingot, 4), 12);
+        TFTreasure.hill1.rare.add(MaterialLibAPI.getStack(Materials.Thaumium, Shapes.dust, 1), 6);
+        TFTreasure.hill1.ultrarare.add(MaterialLibAPI.getStack(Materials.Aluminium, Shapes.dustSmall, 4), 12);
         // Hill2 = medium
-        TFTreasure.hill2.common.add(MaterialLibAPI.getStack(Materials.Thaumium, Shapes.dustTiny, (int) (4L)), 12);
-        TFTreasure.hill2.common.add(MaterialLibAPI.getStack(Materials.Manganese, Shapes.ingot, (int) (4L)), 12);
-        TFTreasure.hill2.common.add(MaterialLibAPI.getStack(Materials.Antimony, Shapes.ingot, (int) (4L)), 12);
-        TFTreasure.hill2.common.add(MaterialLibAPI.getStack(Materials.Bronze, Shapes.ingot, (int) (4L)), 12);
-        TFTreasure.hill2.common.add(MaterialLibAPI.getStack(Materials.Nickel, Shapes.ingot, (int) (4L)), 12);
-        TFTreasure.hill2.common.add(MaterialLibAPI.getStack(Materials.Tin, Shapes.ingot, (int) (6L)), 12);
-        TFTreasure.hill2.common.add(MaterialLibAPI.getStack(Materials.Electrum, Shapes.ingot, (int) (2L)), 6);
-        TFTreasure.hill2.common.add(MaterialLibAPI.getStack(Materials.Copper, Shapes.ingot, (int) (6L)), 12);
-        TFTreasure.hill2.common.add(MaterialLibAPI.getStack(Materials.Magnesium, Shapes.ingot, (int) (4L)), 12);
-        TFTreasure.hill2.uncommon.add(MaterialLibAPI.getStack(Materials.Thaumium, Shapes.dustSmall, (int) (4L)), 12);
+        TFTreasure.hill2.common.add(MaterialLibAPI.getStack(Materials.Thaumium, Shapes.dustTiny, 4), 12);
+        TFTreasure.hill2.common.add(MaterialLibAPI.getStack(Materials.Manganese, Shapes.ingot, 4), 12);
+        TFTreasure.hill2.common.add(MaterialLibAPI.getStack(Materials.Antimony, Shapes.ingot, 4), 12);
+        TFTreasure.hill2.common.add(MaterialLibAPI.getStack(Materials.Bronze, Shapes.ingot, 4), 12);
+        TFTreasure.hill2.common.add(MaterialLibAPI.getStack(Materials.Nickel, Shapes.ingot, 4), 12);
+        TFTreasure.hill2.common.add(MaterialLibAPI.getStack(Materials.Tin, Shapes.ingot, 6), 12);
+        TFTreasure.hill2.common.add(MaterialLibAPI.getStack(Materials.Electrum, Shapes.ingot, 2), 6);
+        TFTreasure.hill2.common.add(MaterialLibAPI.getStack(Materials.Copper, Shapes.ingot, 6), 12);
+        TFTreasure.hill2.common.add(MaterialLibAPI.getStack(Materials.Magnesium, Shapes.ingot, 4), 12);
+        TFTreasure.hill2.uncommon.add(MaterialLibAPI.getStack(Materials.Thaumium, Shapes.dustSmall, 4), 12);
         TFTreasure.hill2.uncommon.add(getModItem(Thaumcraft.ID, "ItemShard", 4, 0), 12);
         TFTreasure.hill2.uncommon.add(getModItem(Thaumcraft.ID, "ItemShard", 4, 1), 12);
         TFTreasure.hill2.uncommon.add(getModItem(Thaumcraft.ID, "ItemShard", 4, 2), 12);
         TFTreasure.hill2.uncommon.add(getModItem(Thaumcraft.ID, "ItemShard", 4, 3), 12);
         TFTreasure.hill2.uncommon.add(getModItem(Thaumcraft.ID, "ItemShard", 4, 4), 12);
         TFTreasure.hill2.uncommon.add(getModItem(Thaumcraft.ID, "ItemShard", 4, 5), 12);
-        TFTreasure.hill2.uncommon.add(MaterialLibAPI.getStack(Materials.DamascusSteel, Shapes.ingot, (int) (4L)), 12);
-        TFTreasure.hill2.uncommon.add(MaterialLibAPI.getStack(Materials.ElectricalSteel, Shapes.ingot, (int) (4L)), 12);
-        TFTreasure.hill2.uncommon.add(MaterialLibAPI.getStack(Materials.Cobalt, Shapes.ingot, (int) (4L)), 12);
-        TFTreasure.hill2.uncommon.add(MaterialLibAPI.getStack(Materials.Molybdenum, Shapes.ingot, (int) (4L)), 12);
-        TFTreasure.hill2.uncommon.add(MaterialLibAPI.getStack(Materials.RedAlloy, Shapes.ingot, (int) (4L)), 12);
-        TFTreasure.hill2.uncommon.add(MaterialLibAPI.getStack(Materials.CastIron, Shapes.ingot, (int) (4L)), 12);
-        TFTreasure.hill2.uncommon.add(MaterialLibAPI.getStack(Materials.Steel, Shapes.ingot, (int) (4L)), 12);
-        TFTreasure.hill2.rare.add(MaterialLibAPI.getStack(Materials.Thaumium, Shapes.dust, (int) (1L)), 12);
-        TFTreasure.hill2.rare.add(MaterialLibAPI.getStack(Materials.Mica, Shapes.dust, (int) (16L)), 12);
-        TFTreasure.hill2.rare.add(MaterialLibAPI.getStack(Materials.Aluminium, Shapes.dust, (int) (2L)), 12);
-        TFTreasure.hill2.ultrarare.add(MaterialLibAPI.getStack(Materials.Gallium, Shapes.ingot, (int) (1L)), 6);
+        TFTreasure.hill2.uncommon.add(MaterialLibAPI.getStack(Materials.DamascusSteel, Shapes.ingot, 4), 12);
+        TFTreasure.hill2.uncommon.add(MaterialLibAPI.getStack(Materials.ElectricalSteel, Shapes.ingot, 4), 12);
+        TFTreasure.hill2.uncommon.add(MaterialLibAPI.getStack(Materials.Cobalt, Shapes.ingot, 4), 12);
+        TFTreasure.hill2.uncommon.add(MaterialLibAPI.getStack(Materials.Molybdenum, Shapes.ingot, 4), 12);
+        TFTreasure.hill2.uncommon.add(MaterialLibAPI.getStack(Materials.RedAlloy, Shapes.ingot, 4), 12);
+        TFTreasure.hill2.uncommon.add(MaterialLibAPI.getStack(Materials.CastIron, Shapes.ingot, 4), 12);
+        TFTreasure.hill2.uncommon.add(MaterialLibAPI.getStack(Materials.Steel, Shapes.ingot, 4), 12);
+        TFTreasure.hill2.rare.add(MaterialLibAPI.getStack(Materials.Thaumium, Shapes.dust, 1), 12);
+        TFTreasure.hill2.rare.add(MaterialLibAPI.getStack(Materials.Mica, Shapes.dust, 16), 12);
+        TFTreasure.hill2.rare.add(MaterialLibAPI.getStack(Materials.Aluminium, Shapes.dust, 2), 12);
+        TFTreasure.hill2.ultrarare.add(MaterialLibAPI.getStack(Materials.Gallium, Shapes.ingot, 1), 6);
         TFTreasure.hill2.ultrarare.add(getModItem(DraconicEvolution.ID, "dezilsMarshmallow", 1, 0), 1);
         // Hill3 = huge
-        TFTreasure.hill3.common.add(MaterialLibAPI.getStack(Materials.Thaumium, Shapes.dustTiny, (int) (6L)), 12);
-        TFTreasure.hill3.common.add(MaterialLibAPI.getStack(Materials.Manganese, Shapes.ingot, (int) (6L)), 12);
-        TFTreasure.hill3.common.add(MaterialLibAPI.getStack(Materials.Antimony, Shapes.ingot, (int) (6L)), 12);
-        TFTreasure.hill3.common.add(MaterialLibAPI.getStack(Materials.Tin, Shapes.ingot, (int) (6L)), 12);
-        TFTreasure.hill3.common.add(MaterialLibAPI.getStack(Materials.Electrum, Shapes.ingot, (int) (2L)), 12);
-        TFTreasure.hill3.common.add(MaterialLibAPI.getStack(Materials.Copper, Shapes.ingot, (int) (6L)), 12);
-        TFTreasure.hill3.common.add(MaterialLibAPI.getStack(Materials.Magnesium, Shapes.ingot, (int) (4L)), 12);
-        TFTreasure.hill3.uncommon.add(MaterialLibAPI.getStack(Materials.Thaumium, Shapes.dustSmall, (int) (6L)), 12);
+        TFTreasure.hill3.common.add(MaterialLibAPI.getStack(Materials.Thaumium, Shapes.dustTiny, 6), 12);
+        TFTreasure.hill3.common.add(MaterialLibAPI.getStack(Materials.Manganese, Shapes.ingot, 6), 12);
+        TFTreasure.hill3.common.add(MaterialLibAPI.getStack(Materials.Antimony, Shapes.ingot, 6), 12);
+        TFTreasure.hill3.common.add(MaterialLibAPI.getStack(Materials.Tin, Shapes.ingot, 6), 12);
+        TFTreasure.hill3.common.add(MaterialLibAPI.getStack(Materials.Electrum, Shapes.ingot, 2), 12);
+        TFTreasure.hill3.common.add(MaterialLibAPI.getStack(Materials.Copper, Shapes.ingot, 6), 12);
+        TFTreasure.hill3.common.add(MaterialLibAPI.getStack(Materials.Magnesium, Shapes.ingot, 4), 12);
+        TFTreasure.hill3.uncommon.add(MaterialLibAPI.getStack(Materials.Thaumium, Shapes.dustSmall, 6), 12);
         TFTreasure.hill3.uncommon.add(getModItem(Thaumcraft.ID, "ItemShard", 4, 0), 12);
         TFTreasure.hill3.uncommon.add(getModItem(Thaumcraft.ID, "ItemShard", 4, 1), 12);
         TFTreasure.hill3.uncommon.add(getModItem(Thaumcraft.ID, "ItemShard", 4, 2), 12);
@@ -74,25 +74,24 @@ public class TF_Loot_Chests {
         TFTreasure.hill3.uncommon.add(getModItem(Thaumcraft.ID, "ItemShard", 4, 4), 12);
         TFTreasure.hill3.uncommon.add(getModItem(Thaumcraft.ID, "ItemShard", 4, 5), 12);
         TFTreasure.hill3.uncommon.add(getModItem(Thaumcraft.ID, "ItemShard", 2, 6), 12);
-        TFTreasure.hill3.uncommon.add(MaterialLibAPI.getStack(Materials.DamascusSteel, Shapes.ingot, (int) (4L)), 12);
-        TFTreasure.hill3.uncommon.add(MaterialLibAPI.getStack(Materials.ElectricalSteel, Shapes.ingot, (int) (4L)), 12);
-        TFTreasure.hill3.uncommon.add(MaterialLibAPI.getStack(Materials.Cobalt, Shapes.ingot, (int) (4L)), 12);
-        TFTreasure.hill3.uncommon.add(MaterialLibAPI.getStack(Materials.Molybdenum, Shapes.ingot, (int) (4L)), 12);
-        TFTreasure.hill3.uncommon.add(MaterialLibAPI.getStack(Materials.RedAlloy, Shapes.ingot, (int) (4L)), 12);
-        TFTreasure.hill3.uncommon.add(MaterialLibAPI.getStack(Materials.CastIron, Shapes.ingot, (int) (4L)), 12);
-        TFTreasure.hill3.uncommon.add(MaterialLibAPI.getStack(Materials.Steel, Shapes.ingot, (int) (4L)), 12);
-        TFTreasure.hill3.rare.add(MaterialLibAPI.getStack(Materials.Thaumium, Shapes.dust, (int) (1L)), 12);
-        TFTreasure.hill3.rare.add(MaterialLibAPI.getStack(Materials.Aluminium, Shapes.dust, (int) (2L)), 12);
-        TFTreasure.hill3.rare.add(MaterialLibAPI.getStack(Materials.Aluminium, Shapes.ingot, (int) (1L)), 12);
-        TFTreasure.hill3.rare.add(MaterialLibAPI.getStack(Materials.Mica, Shapes.dust, (int) (16L)), 12);
+        TFTreasure.hill3.uncommon.add(MaterialLibAPI.getStack(Materials.DamascusSteel, Shapes.ingot, 4), 12);
+        TFTreasure.hill3.uncommon.add(MaterialLibAPI.getStack(Materials.ElectricalSteel, Shapes.ingot, 4), 12);
+        TFTreasure.hill3.uncommon.add(MaterialLibAPI.getStack(Materials.Cobalt, Shapes.ingot, 4), 12);
+        TFTreasure.hill3.uncommon.add(MaterialLibAPI.getStack(Materials.Molybdenum, Shapes.ingot, 4), 12);
+        TFTreasure.hill3.uncommon.add(MaterialLibAPI.getStack(Materials.RedAlloy, Shapes.ingot, 4), 12);
+        TFTreasure.hill3.uncommon.add(MaterialLibAPI.getStack(Materials.CastIron, Shapes.ingot, 4), 12);
+        TFTreasure.hill3.uncommon.add(MaterialLibAPI.getStack(Materials.Steel, Shapes.ingot, 4), 12);
+        TFTreasure.hill3.rare.add(MaterialLibAPI.getStack(Materials.Thaumium, Shapes.dust, 1), 12);
+        TFTreasure.hill3.rare.add(MaterialLibAPI.getStack(Materials.Aluminium, Shapes.dust, 2), 12);
+        TFTreasure.hill3.rare.add(MaterialLibAPI.getStack(Materials.Aluminium, Shapes.ingot, 1), 12);
+        TFTreasure.hill3.rare.add(MaterialLibAPI.getStack(Materials.Mica, Shapes.dust, 16), 12);
         TFTreasure.hill3.rare.add(getModItem(DraconicEvolution.ID, "dezilsMarshmallow", 1, 0), 2);
-        TFTreasure.hill3.ultrarare.add(MaterialLibAPI.getStack(Materials.Gallium, Shapes.ingot, (int) (1L)), 6);
+        TFTreasure.hill3.ultrarare.add(MaterialLibAPI.getStack(Materials.Gallium, Shapes.ingot, 1), 6);
 
         // UrGhast Tower Basment
         TFTreasure.basement.common.add(new ItemStack(Blocks.torch, 16, 0), 2);
-        TFTreasure.basement.common
-                .add(MaterialLibAPI.getStack(Materials.StainlessSteel, Shapes.ingot, (int) (16L)), 12);
-        TFTreasure.basement.common.add(MaterialLibAPI.getStack(Materials.Void, Shapes.ingot, (int) (8L)), 6);
+        TFTreasure.basement.common.add(MaterialLibAPI.getStack(Materials.StainlessSteel, Shapes.ingot, 16), 12);
+        TFTreasure.basement.common.add(MaterialLibAPI.getStack(Materials.Void, Shapes.ingot, 8), 6);
         TFTreasure.basement.uncommon.add(new ItemStack(Items.potionitem, 1, 8198), 12); // potion night vision i
         TFTreasure.basement.uncommon.add(getModItem(Thaumcraft.ID, "ItemResource", 8, 1), 12); // Nitor
         TFTreasure.basement.uncommon.add(new ItemStack(Items.potionitem, 1, 8193), 6); // potion regen i
@@ -125,12 +124,10 @@ public class TF_Loot_Chests {
         TFTreasure.tower_library.common.add(getModItem(Thaumcraft.ID, "blockCandle", 4, 13), 12);
         TFTreasure.tower_library.common.add(getModItem(Thaumcraft.ID, "blockCandle", 4, 14), 12);
         TFTreasure.tower_library.common.add(getModItem(Thaumcraft.ID, "blockCandle", 4, 15), 12);
-        TFTreasure.tower_library.common
-                .add(MaterialLibAPI.getStack(Materials.StainlessSteel, Shapes.dust, (int) (4L)), 6);
+        TFTreasure.tower_library.common.add(MaterialLibAPI.getStack(Materials.StainlessSteel, Shapes.dust, 4), 6);
         TFTreasure.tower_library.common.add(GTOreDictUnificator.get(OrePrefixes.gem, Materials.EnderPearl, 8L), 6);
-        TFTreasure.tower_library.common
-                .add(MaterialLibAPI.getStack(Materials.StainlessSteel, Shapes.ingot, (int) (2L)), 6);
-        TFTreasure.tower_library.common.add(MaterialLibAPI.getStack(Materials.Void, Shapes.dust, (int) (2L)), 6);
+        TFTreasure.tower_library.common.add(MaterialLibAPI.getStack(Materials.StainlessSteel, Shapes.ingot, 2), 6);
+        TFTreasure.tower_library.common.add(MaterialLibAPI.getStack(Materials.Void, Shapes.dust, 2), 6);
         TFTreasure.tower_library.common.add(getModItem(Thaumcraft.ID, "ItemResource", 2, 1), 12); // Nitor
         TFTreasure.tower_library.common.add(new ItemStack(Items.potionitem, 1, 8201), 6); // potion strength i
         TFTreasure.tower_library.common.add(getModItem(Thaumcraft.ID, "ItemResource", 1, 0), 12); // Alumentum
@@ -159,8 +156,7 @@ public class TF_Loot_Chests {
         // TFTreasure.tower_room.common.add(NHItemList.LichBone.get(2),50);
         // TFTreasure.tower_room.common.add(NHItemList.LichBoneChip.get(8),75);
         TFTreasure.tower_room.uncommon.add(getModItem(Thaumcraft.ID, "ItemLootBag", 1, 1), 12);
-        TFTreasure.tower_room.uncommon
-                .add(MaterialLibAPI.getStack(Materials.NetherStar, Shapes.dustTiny, (int) (1L)), 2);
+        TFTreasure.tower_room.uncommon.add(MaterialLibAPI.getStack(Materials.NetherStar, Shapes.dustTiny, 1), 2);
         TFTreasure.tower_room.rare.add(getModItem(Thaumcraft.ID, "ItemLootBag", 1, 2), 6);
     }
 }
